Add comprehensive test coverage for core functionality
High-priority test coverage (47 total tests, +26 new):

## ProcessRunner Tests (11 tests)
- commandExists validation for valid/invalid commands
- run method with exit codes (success, failure)
- Argument passing and working directory support
- Absolute path executable handling
- ProcessError descriptions

## ConfigDiscovery Tests (+5 new, 16 total)
- Directory tree search (current, parent, multiple levels)
- Multiple configs error handling
- Closest config preference in nested directories
- All existing tests continue to pass

## Custom SwiftSyntax Rules Tests (10 tests)
- skimmable_body: 15-line limit for View bodies
- no_group_body: Top-level Group detection (with modifier exception)
- one_top_level_view: Single top-level view enforcement (if-else as one)
- excessive_indentation: 16-space physical indentation limit

## Infrastructure Improvements
- Add Equatable conformance to ConfigDiscoveryError (for test expectations)
- Fix Int32/Int type mismatches in SwiftFormatSmart and SwiftLintCustomSmart
- Add missing findSwiftLintConfig helper function

## Test Results
```
􁁛  Test run with 47 tests in 4 suites passed after 0.041 seconds.
```

All tests passing ✓

+++ Sources/SwiftLintCustomSmart/main.swift
@@ -85,12 +85,12 @@ struct SwiftLintCustomSmart: ParsableCommand {
     }
 
     /// Throw an error if the build failed
-    private func throwIfBuildFailed(_ exitCode: Int) throws {
+    private func throwIfBuildFailed(_ exitCode: Int32) throws {
         guard exitCode == 0 else {
             let errorMsg = ErrorFormatter.formatBuildError(
                 tool: "SwiftLintCustomSmart",
                 project: "rule-engine",
-                exitCode: Int32(exitCode),
+                exitCode: exitCode,
             )
             Console.error(errorMsg)
             throw ExitCode.failure
